+++ src/Copse.Linq.Async/Treenumerables/SelectWhere/SelectWhereComposition.cs
@@ -81,7 +81,7 @@ namespace Copse.Linq.Async.Treenumerables
         var innerResult = innerResultSelector(nodeContext);
 
         return new SelectWhereResult<TOuterResult>(
-          selector(new NodeContext<TResult>(innerResult.Value, nodeContext.Position)),
+          selector(new NodeContext<TResult>(innerResult.Node, nodeContext.Position)),
           innerResult.Strategies);
       };
     }
@@ -97,9 +97,9 @@ namespace Copse.Linq.Async.Treenumerables
         var innerResult = innerResultSelector(nodeContext);
 
         return new SelectWhereResult<TResult>(
-          innerResult.Value,
+          innerResult.Node,
           innerResult.Strategies
-            | (predicate(new NodeContext<TResult>(innerResult.Value, nodeContext.Position))
+            | (predicate(new NodeContext<TResult>(innerResult.Node, nodeContext.Position))
               ? NodeTraversalStrategies.SkipDescendants
               : NodeTraversalStrategies.TraverseAll));
       };