+++ src/Copse.Linq.Async/Treenumerators/Filter/AsyncScanWhereDepthFirstTreenumerator.cs
@@ -144,7 +144,7 @@ namespace Copse.Linq.Async
 
       var cacheChild = _Path.ShouldCacheChild();
 
-      _Path.PushAcceptedChild(result.Value, InnerTreenumerator.Position);
+      _Path.PushAcceptedChild(result.Node, InnerTreenumerator.Position);
 
       if (cacheChild)
       {