+++ src/Copse.Linq/Treenumerators/Filter/ScanWhereBreadthFirstTreenumerator.g.cs
@@ -279,7 +279,7 @@ namespace Copse.Linq.Treenumerators
             if (!_Path.ConsumerSkippedChildAfterLastAccepted)
             {
               _Path.ClearConsumerSkippedChildAfterLastAccepted();
-              _Path.EnqueueAccepted(result.Value, effectivePosition, innerDepth);
+              _Path.EnqueueAccepted(result.Node, effectivePosition, innerDepth);
               _Path.MarkDeferredSchedulePending();
               _DeferredResultStrategies = result.Strategies;
               _Path.Front.VisitCount++;
@@ -291,7 +291,7 @@ namespace Copse.Linq.Treenumerators
             _Path.ClearConsumerSkippedChildAfterLastAccepted();
           }
 
-          _Path.EnqueueAccepted(result.Value, effectivePosition, innerDepth);
+          _Path.EnqueueAccepted(result.Node, effectivePosition, innerDepth);
           _PendingResultStrategies = result.Strategies;
         }
         else // VisitingNode